refactor: move home components into own page folder
